enterprise-rag-platform/requirements.txt
pypdf>=4.0
chromadb>=0.4.22
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
openai>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
